Add support for deployment via Juju Charms
This merge request contains the source code to the layers and interfaces required to build the charms that will deploy the OpenMANO stack, with OpenVIM, in a cloud via Juju.

There also contains a number of fixes made by our team in order to get OpenVim running on Juju, but were made prior to the fork from the openmano project. As such, those changes should be reviewed more carefully.